Rename and move `concurrent.WaitFor` to `common.WaitForValue`
It seems that `WaitFor` function doesn't relate to concurrent package.
Also added a new version of `WaitFor` which doesn't return value.
Actually, we do not need to return any value because we always can use
clojures to pass values from the function body.
