+++ nil/common/concurrent/utils.go
@@ -41,36 +41,6 @@ func RunWithTimeout(ctx context.Context, timeout time.Duration, fs ...Func) erro
 	return nil
 }
 
-// WaitFor repeatedly calls the given function until it returns true or an error.
-// In case function return some data not equal to nil, it returns true.
-func WaitFor[T any](
-	ctx context.Context,
-	timeout,
-	tick time.Duration,
-	f func(ctx context.Context) (*T, error),
-) (*T, error) {
-	ctx, cancel := context.WithTimeout(ctx, timeout)
-	defer cancel()
-
-	ticker := time.NewTicker(tick)
-	defer ticker.Stop()
-
-	for {
-		select {
-		case <-ctx.Done():
-			return nil, ctx.Err()
-		case <-ticker.C:
-			res, err := f(ctx)
-			if err != nil {
-				return res, err
-			}
-			if res != nil {
-				return res, nil
-			}
-		}
-	}
-}
-
 // Run calls RunWithTimeout without a timeout.
 func Run(ctx context.Context, fs ...Func) error {
 	return RunWithTimeout(ctx, 0, fs...)
